Add error handling, refactor search as components
Added error text that appears when the user enters
a city that can't be found by Google.

Added functionality to remove suggestions when Google
places returns a 'ZERO_RESULTS' error.

Moved RequestLocation and SearchLocation to individual
components, they each generate the weather data and then
pass it back to App.js to update state and WeatherStatus